+++ src/designs/basic/screens/welcome-screen/WelcomeScreen.tsx
